backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
twilio==9.3.7
anthropic==0.39.0
httpx==0.27.2
openai==1.54.5
python-dotenv==1.0.1
aiosqlite==0.20.0
pydantic==2.9.2
